+++ backend/models.go
@@ -22,42 +22,42 @@ type ReceiptResponse struct {
 type Item struct {
 	Name        string  `json:"name"`
 	NDS         int     `json:"nds"`
-	NDSSum      float64 `json:"ndsSum"`
+	NDSSum      float32 `json:"ndsSum"`
 	PaymentType int     `json:"paymentType"`
-	Price       float64 `json:"price"`
+	Price       float32 `json:"price"`
 	ProductType int     `json:"productType"`
 	ProviderINN string  `json:"providerInn"`
-	Quantity    float64 `json:"quantity"`
-	Sum         float64 `json:"sum"`
+	Quantity    float32 `json:"quantity"`
+	Sum         float32 `json:"sum"`
 }
 
 type FiscalDataResponse struct {
 	BuyerPhoneOrAddress     string  `json:"buyerAddress"`
-	CashTotalSum            float64 `json:"cashTotalSum"`
+	CashTotalSum            float32 `json:"cashTotalSum"`
 	Code                    int     `json:"code"`
-	CreditSum               float64 `json:"creditSum"`
+	CreditSum               float32 `json:"creditSum"`
 	DateTime                string  `json:"dateTime"`
-	ECashTotalSum           float64 `json:"ecashTotalSum"`
+	ECashTotalSum           float32 `json:"ecashTotalSum"`
 	FiscalDocumentFormatVer string  `json:"fiscalDocumentFormatVer"`
 	FiscalDocumentNumber    int     `json:"fiscalDocumentNumber"`
 	FiscalDriveNumber       string  `json:"fiscalDriveNumber"`
 	FiscalSign              string  `json:"fiscalSign"`
 	Items                   []Item  `json:"items"`
 	KKTRegID                string  `json:"kktRegId"`
-	NDS10                   float64 `json:"nds10"`
-	NDS18                   float64 `json:"nds18"`
+	NDS10                   float32 `json:"nds10"`
+	NDS18                   float32 `json:"nds18"`
 	OperationType           int     `json:"operationType"`
 	Operator                string  `json:"operator"`
 	OperatorINN             string  `json:"operatorInn"`
-	PrepaidSum              float64 `json:"prepaidSum"`
-	ProvisionSum            float64 `json:"provisionSum"`
+	PrepaidSum              float32 `json:"prepaidSum"`
+	ProvisionSum            float32 `json:"provisionSum"`
 	RequestNumber           int     `json:"requestNumber"`
 	RetailPlace             string  `json:"retailPlace"`
 	RetailPlaceAddress      string  `json:"retailPlaceAddress"`
 	ShiftNumber             int     `json:"shiftNumber"`
 	TaxationType            int     `json:"taxationType"`
 	AppliedTaxationType     int     `json:"appliedTaxationType"`
-	TotalSum                float64 `json:"totalSum"`
+	TotalSum                float32 `json:"totalSum"`
 	User                    string  `json:"user"`
 	UserINN                 string  `json:"userInn"`
 }
